segmenter/pyproject.toml
[tool.poetry]
name = "segmenter"
version = "0.1.25"
description = ""
authors = ["Antoine Vandermeersch <avdmeers@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.7"
boto3 = "^1.28.84"
supervisor = "^4.2.5"
click-logging = "^1.0.1"
pydantic = "^2.5.1"
pillow = "^10.1.0"

[tool.poetry.scripts]
segmenter = "segmenter.console:main"

[tool.poetry.group.dev.dependencies]
ipython = "^8.17.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
